@@ -64,9 +64,11 @@ protected:
   /// Get the list of hadrons for the upcoming timestep from BulkDynamicsManager (will clear the list)
   std::vector<std::shared_ptr<Hadron>> GetTimestepParticlizationHadrons();
 
+  std::vector<std::vector<std::shared_ptr<Hadron>>> dummy;
+  std::uniform_real_distribution<double> ZeroOneDistribution;
+
   // rng for the Kaon-L / Kaon-S switch to K0 / Anti-K0
   std::shared_ptr<std::uniform_int_distribution<int>> rand_int_ptr_;
-
 };
 
 } // end namespace Jetscape
