@@ -323,7 +323,18 @@ void JetScape::DetermineTaskListFromXML() {
             JSINFO << "JetScape::DetermineTaskList() -- Hard Process: Added "
                       "PythiaGun to task list.";
           }
-        } else if (((int)childElementName.find("CustomModule") >= 0)) {
+        }
+        //   - epemGun
+        else if (childElementName == "epemGun") {
+          auto EpemGun =
+              JetScapeModuleFactory::createInstance(childElementName);
+          if (EpemGun) {
+            Add(EpemGun);
+            JSINFO << "JetScape::DetermineTaskList() -- Hard Process: Added "
+                      "epemGun to task list.";
+          }
+        }
+        else if (((int)childElementName.find("CustomModule") >= 0)) {
           auto customModule =
               JetScapeModuleFactory::createInstance(childElementName);
           if (customModule) {
@@ -916,6 +927,10 @@ void JetScape::SetPointers() {
       JetScapeSignalManager::Instance()->SetSoftParticlizationPointer(
           dynamic_pointer_cast<SoftParticlization>(it));
       iss_pointer_is_set = true;
+      JetScapeSignalManager::Instance()->ConnectGetHydroHyperSurfaceSignal(
+          dynamic_pointer_cast<SoftParticlization>(it));
+      JetScapeSignalManager::Instance()->ConnectClearHydroHyperSurfaceSignal(
+          dynamic_pointer_cast<SoftParticlization>(it));
     } else if (dynamic_pointer_cast<HadronizationManager>(it)) {
       JetScapeSignalManager::Instance()->SetHadronizationManagerPointer(
 										dynamic_pointer_cast<HadronizationManager>(it));
