@@ -28,7 +28,6 @@ Pythia8::Pythia PGun::InternalHelperPythia("IntentionallyEmpty", false);
 PGun::PGun() : HardProcess() {
   fixed_pT = 0;
   parID = 21;
-  flag_useHybridHad = 0;
   SetId("PGun");
   VERBOSE(8);
 }
@@ -48,9 +47,6 @@ void PGun::InitTask() {
 
   parID = GetXMLElementDouble({"Hard", "PGun", "parID"});
   JSINFO << "Parton Gun with parID = " << parID;
-
-  flag_useHybridHad = GetXMLElementInt({"Hard", "PGun", "useHybridHad"});
-  JSINFO << "Use hybrid hadronization? " << flag_useHybridHad;
 }
 
 void PGun::ExecuteTask() {
@@ -58,7 +54,7 @@ void PGun::ExecuteTask() {
 
   double p[4], xLoc[4];
 
-  double pT, rapidity, phi;
+  double pT, rapidity, phi, pseudorapidity;
   double eta_cut = 1.0;
   double tempRand;
   const double maxN = 1.0 * RAND_MAX;
@@ -91,6 +87,13 @@ void PGun::ExecuteTask() {
   p[3] = sqrt(pT * pT + mass * mass) * sinh(rapidity);
   p[0] = sqrt(pT * pT + mass * mass) * cosh(rapidity);
 
+  double p_abs = std::sqrt(pT*pT + p[3]*p[3]);
+  if(std::abs(p_abs - p[3]) > rounding_error){
+    pseudorapidity = 0.5 * std::log((p_abs + p[3]) / (p_abs - p[3]));
+  } else {
+    JSWARN << "Particle in PGun has infinite pseudorapidity.";
+  }
+
   // Roll for a starting point
   // See: https://stackoverflow.com/questions/15039688/random-generator-from-vector-with-probability-distribution-in-c
   for (int i = 0; i <= 3; i++) {
@@ -114,18 +117,11 @@ void PGun::ExecuteTask() {
   xLoc[2] = 0.0;
   xLoc[3] = 0.0;
   
-  if (flag_useHybridHad != 1) {
-    AddParton(make_shared<Parton>(0, parID, 0, pT, rapidity, phi, p[0], xLoc));
-  } else {
-    auto ptn = make_shared<Parton>(0, parID, 0, pT, rapidity, phi, p[0], xLoc);
-    ptn->set_color((parID > 0) ? 100 : 0);
-    ptn->set_anti_color(((parID > 0) || (parID == 21)) ? 0 : 101);
-    ptn->set_max_color(102);
-    AddParton(ptn);
-  }
-
-  //      VERBOSEPARTON(7,*GetPartonAt(i)) <<" added "<<" at x=" << xLoc[1]<<", y=" << xLoc[2]<<", z=" << xLoc[3];
-  //    }
+  auto ptn = make_shared<Parton>(0, parID, 0, pT, pseudorapidity, phi, p[0], xLoc);
+  ptn->set_color((parID > 0) ? 100 : 0);
+  ptn->set_anti_color(((parID > 0) || (parID == 21)) ? 0 : 101);
+  ptn->set_max_color(102);
+  AddParton(ptn);
 
   VERBOSE(8) << GetNHardPartons();
 }
